Fix: 위경도값 자료형 변경(double to BigDecimal)

+++ src/main/java/com/gdsc/plogger/trashSpot/data/dto/res/TrashSpotGetRes.java
@@ -1,13 +1,14 @@
 package com.gdsc.plogger.trashSpot.data.dto.res;
 
 import com.gdsc.plogger.trashSpot.data.TrashSpot;
+import java.math.BigDecimal;
 import lombok.Data;
 
 @Data
 public class TrashSpotGetRes {
     private Long id;
-    private double latitude;
-    private double longitude;
+    private BigDecimal latitude;
+    private BigDecimal longitude;
 
     public TrashSpotGetRes(TrashSpot trashSpot) {
         this.id = trashSpot.getId();
