+++ src/main/java/com/gdsc/plogger/trashcan/data/Trashcan.java
@@ -4,6 +4,7 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import java.math.BigDecimal;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -19,15 +20,15 @@ public class Trashcan {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private double latitude;
+    private BigDecimal latitude;
 
-    private double longitude;
+    private BigDecimal longitude;
 
     @ColumnDefault("0")
     private int report;
 
     @Builder
-    public Trashcan(double latitude, double longitude) {
+    public Trashcan(BigDecimal latitude, BigDecimal longitude) {
         this.latitude = latitude;
         this.longitude = longitude;
     }
